refactor: remove BatchNLPModels submodule and AbstractExaModel
- Delete src/BatchNLPModels.jl; move FlatNLPModel and generic get_nbatch
  overloads into utils.jl
- Remove abstract type AbstractExaModel; ExaModel now directly subtypes
  NLPModels.AbstractNLPModel
- Update ExaModelsKernelAbstractions.jl and test/BatchTest accordingly
- Remove BatchNLPModels section from docs/src/core.md
