+++ docs/src/core.md
@@ -3,7 +3,3 @@
 Modules = [ExaModels]
 ```
 
-# Batch NLP Models
-```@autodocs
-Modules = [ExaModels.BatchNLPModels]
-```
