+++ ext/ExaModelsKernelAbstractions.jl
@@ -138,7 +138,7 @@ function _grad_structure!(T, backend, (obj, objs...), gsparsity)
 end
 
 function ExaModels.jac_structure!(
-    m::ExaModels.AbstractExaModel{T,VT,E},
+    m::ExaModels.ExaModel{T,VT,E},
     rows::V,
     cols::V,
 ) where {T,VT,E<:KAExtension,V<:AbstractVector}
@@ -155,7 +155,7 @@ end
 
 
 function ExaModels.hess_structure!(
-    m::ExaModels.AbstractExaModel{T,VT,E},
+    m::ExaModels.ExaModel{T,VT,E},
     rows::V,
     cols::V,
 ) where {T,VT,E<:KAExtension,V<:AbstractVector}
@@ -179,7 +179,7 @@ end
 
 
 function ExaModels.obj(
-    m::ExaModels.AbstractExaModel{T,VT,E},
+    m::ExaModels.ExaModel{T,VT,E},
     x::AbstractVector,
 ) where {T,VT,E<:KAExtension}
     if !isempty(m.ext.objbuffer)
@@ -213,7 +213,7 @@ end
 
 
 function ExaModels.cons_nln!(
-    m::ExaModels.AbstractExaModel{T,VT,E},
+    m::ExaModels.ExaModel{T,VT,E},
     x::AbstractVector,
     y::AbstractVector,
 ) where {T,VT,E<:KAExtension}
@@ -250,7 +250,7 @@ function _conaugs!(backend, y, (con, cons...), x, θ)
 end
 
 function ExaModels.grad!(
-    m::ExaModels.AbstractExaModel{T,VT,E},
+    m::ExaModels.ExaModel{T,VT,E},
     x::V,
     y::V,
 ) where {T,VT,E<:KAExtension,V<:AbstractVector}
@@ -278,7 +278,7 @@ function _grad!(backend, y, (obj, objs...), x, θ)
 end
 
 function ExaModels.jac_coord!(
-    m::ExaModels.AbstractExaModel{T,VT,E},
+    m::ExaModels.ExaModel{T,VT,E},
     x::V,
     y::V,
 ) where {T,VT,E<:KAExtension,V<:AbstractVector}
@@ -293,7 +293,7 @@ function _jac_coord!(backend, y, (con, cons...), x, θ)
 end
 
 function ExaModels.jprod_nln!(
-    m::ExaModels.AbstractExaModel{T,VT,E},
+    m::ExaModels.ExaModel{T,VT,E},
     x::AbstractVector,
     v::AbstractVector,
     Jv::AbstractVector,
@@ -301,7 +301,7 @@ function ExaModels.jprod_nln!(
     error("Prodhelper is not defined. Use ExaModels(c; prod=true) to use jprod_nln!")
 end
 function ExaModels.jtprod_nln!(
-    m::ExaModels.AbstractExaModel{T,VT,E},
+    m::ExaModels.ExaModel{T,VT,E},
     x::AbstractVector,
     v::AbstractVector,
     Jtv::AbstractVector,
@@ -309,7 +309,7 @@ function ExaModels.jtprod_nln!(
     error("Prodhelper is not defined. Use ExaModels(c; prod=true) to use jtprod_nln!")
 end
 function ExaModels.jprod_nln!(
-    m::ExaModels.AbstractExaModel{T,VT,E},
+    m::ExaModels.ExaModel{T,VT,E},
     x::AbstractVector,
     v::AbstractVector,
     Jv::AbstractVector,
@@ -329,7 +329,7 @@ function ExaModels.jprod_nln!(
     return Jv
 end
 function ExaModels.jtprod_nln!(
-    m::ExaModels.AbstractExaModel{T,VT,E},
+    m::ExaModels.ExaModel{T,VT,E},
     x::AbstractVector,
     v::AbstractVector,
     Jtv::AbstractVector,
@@ -349,7 +349,7 @@ function ExaModels.jtprod_nln!(
     return Jtv
 end
 function ExaModels.hprod!(
-    m::ExaModels.AbstractExaModel{T,VT,E},
+    m::ExaModels.ExaModel{T,VT,E},
     x::AbstractVector,
     y::AbstractVector,
     v::AbstractVector,
@@ -386,7 +386,7 @@ function ExaModels.hprod!(
     return Hv
 end
 function ExaModels.hprod!(
-    m::ExaModels.AbstractExaModel{T,VT,E},
+    m::ExaModels.ExaModel{T,VT,E},
     x::AbstractVector,
     v::AbstractVector,
     Hv::AbstractVector;
@@ -455,7 +455,7 @@ end
 
 
 function ExaModels.hess_coord!(
-    m::ExaModels.AbstractExaModel{T,VT,E},
+    m::ExaModels.ExaModel{T,VT,E},
     x::V,
     y::V,
     hess::V;
@@ -514,7 +514,7 @@ end
     end
 end
 
-ExaModels.getbackend(m::ExaModels.AbstractExaModel{T,VT,E}) where {T,VT,E<:KAExtension} =
+ExaModels.getbackend(m::ExaModels.ExaModel{T,VT,E}) where {T,VT,E<:KAExtension} =
     m.ext.backend
 function ExaModels._compress!(V, buffer, ptr, sparsity, backend)
     fill!(V, zero(eltype(V)))
